% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inflection_Finder.R
\name{Inflection_Finder}
\alias{Inflection_Finder}
\title{Inflection finder, findds the infection of the emergence curve}
\usage{
Inflection_Finder(list_dynamics)
}
\arguments{
\item{list_dynamics}{Output from the SingleSpecies_Dynamics}
}
\value{
dataframe containing the shape parameter as well as the time of inflection
}
\description{
Inflection finder, findds the infection of the emergence curve
}
\examples{
Inflection_Finder(x)
}
