% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_creator.R
\name{params_creator}
\alias{params_creator}
\title{Creation of parameters to simulate the phenology with}
\usage{
params_creator(
  shape,
  initial_val = 1000,
  dev_val = 1/30,
  mu_val = 0.33,
  m1_val = 8,
  m2_val = 0.5,
  mk_val = 5,
  delay1 = 1/5
)
}
\arguments{
\item{shape}{The shape parameter for the Erlang distribution}

\item{initial_val}{The initial starting value of the first stage individuals (i.e Juveniles)}

\item{dev_val}{The development rate (i.e The mean development time is 1/dev)}

\item{mu_val}{The natural moratlity rate of the second stage individuals (adults)}

\item{m1_val}{The Hill function of the pesticide parameter}

\item{m2_val}{The Hill function of the pesticide parameter}

\item{mk_val}{The Hill function of the pesticide parameter}

\item{delay1}{The decay rate of the pesticide}
}
\value{
A list with two elements: the initial value or the parameters
}
\description{
Creation of parameters to simulate the phenology with
}
\examples{
params_creator(100)
}
