% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_Species_ActionThreshold.R
\name{Single_Species_ActionThreshold}
\alias{Single_Species_ActionThreshold}
\title{Calculate the action threshold for the single species due to the inflection
point}
\usage{
Single_Species_ActionThreshold(output_dataframe, inflection_df)
}
\arguments{
\item{output_dataframe}{The dynamics output}

\item{inflection_df}{The data.frame containing the timing of the inflection
for each of the shape parameter}
}
\value{
A list of the shape and action thresholds
}
\description{
Calculate the action threshold for the single species due to the inflection
point
}
\examples{
Single_Species_ActionThreshold(DF, Inflection_DF)

}
