% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpecies_AUC.R
\name{SingleSpecies_AUC}
\alias{SingleSpecies_AUC}
\title{Calculate the total area under curve without intervention}
\usage{
SingleSpecies_AUC(initial, dev, mu)
}
\arguments{
\item{dev}{The development rate (i.e The mean development time is 1/dev)}

\item{mu}{The natural moratlity rate of the second stage individuals (adults)}

\item{initial_individuals}{The initial starting value of the first stage individuals (i.e Juveniles)}
}
\value{
A data.frame with the starting initial value in the first stage depending on the shape
}
\description{
Due to the attrition in the second stage, the total damage (the area under the curve)
may be different depending on the shape parameter. This function calculates the initial
starting value so that the total insect damage is the same for all shapes IF there is no
interventions
}
\examples{
onesp_insect_damage(1000, 1 / 30, 4)
}
