% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpecies_Dynamics.R
\name{SingleSpecies_Dynamics}
\alias{SingleSpecies_Dynamics}
\title{Plotting the dynamics of a single species and calculating the first and
second derivative}
\usage{
SingleSpecies_Dynamics(list_params_init)
}
\arguments{
\item{list_params_init}{Output from the params_creator (list with two elments)}

\item{plot}{Plot the output as well as the first and second derivatives}
}
\value{
A data.frame of all the results as well as the plots
}
\description{
Plotting the dynamics of a single species and calculating the first and
second derivative
}
\examples{
SingleSpecies_Dynamics(x)
}
