% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singe_species_sprays.R
\name{ss_shape_spray_dynamics}
\alias{ss_shape_spray_dynamics}
\title{Model the pesticide spraying of a single species emergence as well as determining
the number of sprays needed}
\usage{
ss_shape_spray_dynamics(list_params_init, AT)
}
\arguments{
\item{list_params_init}{The parameters generated by the params_creator (a list with two elements)
including the initial conditions and parameters}

\item{AT}{The action threshold to spray for each shape}
}
\value{
A dataframe that contains the outputs as well as the number of spray times
}
\description{
Model the pesticide spraying of a single species emergence as well as determining
the number of sprays needed
}
\examples{
ss_shape_spray_dynamics(list_params, 100)
}
